add removal of tmp file at exit and checking for tmp file to reload

+++ dzen2/reload_dzen.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 # pkill dzen2 ;
 # ~/Applications/openbox-config/dzen2/dzen2_statusbar.sh &
-while read -r var
-do
-  kill -SIGUSR1 $var
-done < "/tmp/dzenpid"
+if [ -f /tmp/dzenpid ]
+then
+  while read -r var
+  do
+    kill -SIGUSR1 $var
+  done < "/tmp/dzenpid"
+fi
